removed secure headers to a different file
