+++ src/zenml/zen_server/utils.py
@@ -67,7 +67,6 @@ _rbac: Optional[RBACInterface] = None
 _feature_gate: Optional[FeatureGateInterface] = None
 _workload_manager: Optional[WorkloadManagerInterface] = None
 _plugin_flavor_registry: Optional[PluginFlavorRegistry] = None
-_secure_headers: Optional[Any] = None
 
 
 def zen_store() -> "SqlZenStore":
@@ -221,105 +220,6 @@ def initialize_zen_store() -> None:
 
     global _zen_store
     _zen_store = zen_store_
-
-
-def secure_headers() -> Any:
-    """Return the secure headers component.
-
-    Returns:
-        The secure headers component.
-
-    Raises:
-        RuntimeError: If the secure headers component is not initialized.
-    """
-    global _secure_headers
-    if _secure_headers is None:
-        raise RuntimeError("Secure headers component not initialized")
-    return _secure_headers
-
-
-def initialize_secure_headers() -> None:
-    """Initialize the secure headers component."""
-    global _secure_headers
-
-    config = server_config()
-
-    # For each of the secure headers supported by the `secure` library, we
-    # check if the corresponding configuration is set in the server
-    # configuration:
-    #
-    # - if set to `True`, we use the default value for the header
-    # - if set to a string, we use the string as the value for the header
-    # - if set to `False`, we don't set the header
-    import secure
-
-    server: Optional[secure.Server] = None
-    if config.secure_headers_server:
-        server = secure.Server()
-        if isinstance(config.secure_headers_server, str):
-            server.set(config.secure_headers_server)
-        else:
-            server.set(str(config.deployment_id))
-
-    hsts: Optional[secure.StrictTransportSecurity] = None
-    if config.secure_headers_hsts:
-        hsts = secure.StrictTransportSecurity()
-        if isinstance(config.secure_headers_hsts, str):
-            hsts.set(config.secure_headers_hsts)
-
-    xfo: Optional[secure.XFrameOptions] = None
-    if config.secure_headers_xfo:
-        xfo = secure.XFrameOptions()
-        if isinstance(config.secure_headers_xfo, str):
-            xfo.set(config.secure_headers_xfo)
-
-    xxp: Optional[secure.XXSSProtection] = None
-    if config.secure_headers_xxp:
-        xxp = secure.XXSSProtection()
-        if isinstance(config.secure_headers_xxp, str):
-            xxp.set(config.secure_headers_xxp)
-
-    csp: Optional[secure.ContentSecurityPolicy] = None
-    if config.secure_headers_csp:
-        csp = secure.ContentSecurityPolicy()
-        if isinstance(config.secure_headers_csp, str):
-            csp.set(config.secure_headers_csp)
-
-    content: Optional[secure.XContentTypeOptions] = None
-    if config.secure_headers_content:
-        content = secure.XContentTypeOptions()
-        if isinstance(config.secure_headers_content, str):
-            content.set(config.secure_headers_content)
-
-    referrer: Optional[secure.ReferrerPolicy] = None
-    if config.secure_headers_referrer:
-        referrer = secure.ReferrerPolicy()
-        if isinstance(config.secure_headers_referrer, str):
-            referrer.set(config.secure_headers_referrer)
-
-    cache: Optional[secure.CacheControl] = None
-    if config.secure_headers_cache:
-        cache = secure.CacheControl()
-        if isinstance(config.secure_headers_cache, str):
-            cache.set(config.secure_headers_cache)
-
-    permissions: Optional[secure.PermissionsPolicy] = None
-    if config.secure_headers_permissions:
-        permissions = secure.PermissionsPolicy()
-        if isinstance(config.secure_headers_permissions, str):
-            permissions.value = config.secure_headers_permissions
-
-    _secure_headers = secure.Secure(
-        server=server,
-        hsts=hsts,
-        xfo=xfo,
-        xxp=xxp,
-        csp=csp,
-        content=content,
-        referrer=referrer,
-        cache=cache,
-        permissions=permissions,
-    )
 
 
 _server_config: Optional[ServerConfiguration] = None
